requirements_alfworld.txt
ai2thor==2.1.0
alfworld==0.3.5
annotated-types==0.7.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
confection==0.1.5
cymem==2.0.8
fast_downward_textworld==20.6.2
filelock==3.16.1
Flask==2.1.3
fsspec==2024.10.0
gym==0.26.2
gym-notices==0.0.8
hashids==1.3.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
jericho==3.3.0
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mementos==1.3.1
more-itertools==10.5.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
preshed==3.0.9
progressbar2==4.5.0
prompt_toolkit==3.0.48
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-utils==3.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.1
TatSu==5.8.3
termcolor==2.5.0
textworld==1.6.1
thinc==8.3.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
triton==3.1.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==2.0.3
wrapt==1.16.0
zipp==3.20.2
